Fix MSVC warning C4018 for sqlite.
This is a warning about signed vs. unsigned values in a comparison.  The code in
question was added entirely by a previous patch, so this simply modifies that
patch, and the results of applying it.

Adding an "(unsigned)" C-style cast is a somewhat ugly fix to one location, but
the alternative would be a potentially-vast change to the APIs in question to
change int to unsigned throughout the transitive closure of places that touch
this.  I didn't seriously investigate doing this.

BUG=398202
TEST=none

Review URL: https://codereview.chromium.org/1248763003

Cr-Original-Commit-Position: refs/heads/master@{#339765}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b0bf37632646237f4005203cf08e76e3b94e545c