Add Chrome OS's minigbm library to third_party
This library is used by Chrome on Chrome OS to allocate accelerated
graphics buffers. The version avaiable on Ubuntu is too old, and as a
result we can't run unit tests for the gbm platform code on developer
workstations or on the waterfall.

This adds the library to third_party and fixes up the build files to use
either the OS copy (for Chrome OS release builds) or gardened copy (for
testing on the waterfall) depending on the value of "use_system_minigbm".

BUG=530712
TEST=
  GYP_DEFINES="use_ozone=1 chromeos=1 ozone_platform_gbm=1 use_system_minigbm=0" gclient runhooks
  ninja -C out/Release ozone_unittests
  ./out/Release/ozone_unittests

  gn gen out_gn_ozone --args='target_os="chromeos" use_ozone=true ozone_platform_gbm=true use_system_minigbm=false'
  ninja -C out_gn_ozone ozone_unittests
  ./out_gn_ozone/Release/ozone_unittests

Review URL: https://codereview.chromium.org/1341473003

Cr-Original-Commit-Position: refs/heads/master@{#348758}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c231bffd37754a1ab09fbbb4a1260438be638df4