[Reland] Invalidate whitespace text on text color changes
This patch updates the optimization added in [1] to invalidate
whitespace text on color changes because whitespace text can have
decorations such as underlines which still need to repaint.

The big idea of the optimization in [1] is to not unnecessarily repaint
expensive items such as images that are in an <a> element. On hover, <a>
will change color and underline styles which should not affect images.
This patch adds a test to prove this optimization still works and
prevents it from breaking in the future.

This patch contains a fix for a performance regression[2] where text
nodes without text (i.e., !firstTextBox) would be invalidated. This is
tested by fast/repaint/inline-outline-repaint-2.html but was incorrectly
thought to be a benign change in the original patch. Additional comments
have been added both about the optimization and about the tricky
empty-text case. Additionally, hasTextBoxes has been introduced to make
this area cleaner.

[1] https://src.chromium.org/viewvc/blink?view=rev&revision=152046
[2] https://src.chromium.org/viewvc/blink?view=rev&revision=202374

TBR=chrishtr
BUG=529944

Review URL: https://codereview.chromium.org/1363913004

Cr-Original-Commit-Position: refs/heads/master@{#350412}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 72f85dacbbef4c117efdb8e353036873d167b989