bluetooth: WebBluetoothGATTCharacteristic -> WebBluetoothGATTCharacteristicInit
We need to be able to communicate BluetoothGATTCharacteristic events from the
embedder. In order to do that we need a new interface. Rename the WebBluetoothGATTCharacteristic struct to
WebBluetoothGATTCharacteristicInit so we can later introduce the class
WebBluetoothGATTCharacteristic that will serve to communicate with the
embedder.

BUG=533713

Review URL: https://codereview.chromium.org/1366653005

Cr-Original-Commit-Position: refs/heads/master@{#350487}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c0cac51602a24ab4c344953b731c48e303690bbd