blink_perf.dom.textarea-dom: vary text nodes added.
This benchmark is meant to measure the performance of a sequence of
text node appends to a <textarea/> element. Have that sequence consist
of varying strings rather than the exact same one. This is done so as to
have the measurement not be influenced by whatever reuse of text
node strings that the underlying allocator is able to provide if using
identical strings.

R=
BUG=

Review URL: https://codereview.chromium.org/1367923004

Cr-Original-Commit-Position: refs/heads/master@{#350533}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8ac201565c743bdf4723d0a4a4061cc3579f9378