Pass correct page URL for context menus
A regression erroneously caused a blank page URL to be passed in the
context menu creation parameters from the renderer process to the
browser, if the menu was created in a subframe.

This corrects the problem, and allows context menus to properly account
for having remote main frames in site per process mode.

BUG=534209
R=dcheng@chromium.org

Review URL: https://codereview.chromium.org/1359693003

Cr-Original-Commit-Position: refs/heads/master@{#350595}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 60a9983a9d4b91c8c74174acc563fd1b65932b3b