Add some details about the expected table structure
This change explains the structure LayoutTable and all the table parts
are expecting. It adds an example where this yield to some complexity.

Also to be able to explain this correctly, I added a lot of
explanations on anonymous wrapper generation in LayoutObject.

Review URL: https://codereview.chromium.org/1362273003

Cr-Original-Commit-Position: refs/heads/master@{#350781}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4f1e1d77745edb9a1c2f9c52c8c04e0b62b21017