Allocate TransformationMatrix on PartitionAlloc
TransformationMatrix needs to be 16-byte aligned on 64-bit platforms
due to use of vector instructions. PartitionAlloc has an ability to allocate
a multiple-of-16-byte object with 16-byte alignment, but Oilpan doesn't have
the ability.

So this CL changes CSSMatrix and DOMMatrixReadOnly so that they hold
TransformationMatrix via an OwnPtr, which makes sure that TransformationMatrix
is allocated on PartitionAlloc.

Patch by haraken@chromium.org, reviewed at
https://codereview.chromium.org/1361373002/
This is identical to this patch, except that the check in
checkAlignment() is done conditionally.

BUG=534853
TBR=haraken

Review URL: https://codereview.chromium.org/1359363005

Cr-Original-Commit-Position: refs/heads/master@{#350931}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 997afddb1a38bba06bfd6f048b207434aebc967a