When style changes, set the LayoutObject as a client for all StyleImage loads
We have logic to dynamically change an image load's
ResourceLoadPriority based on whether it is in the viewport. We would like
to generalize this logic to support other types of resources and other
factors for priority. To do that, we need to ensure every image that might
be in the viewport is associated with a LayoutObject, so that it can pull its
priority state, rather than having the LayoutObject push the state.

We have not been consistent about ensuring that every LayoutObject registers
itself as a client with every StyleImage in its ComputedStyle. This change
ensures that we do so.

BUG=536147

Review URL: https://codereview.chromium.org/1362693004

Cr-Original-Commit-Position: refs/heads/master@{#350946}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c5ce89c70400c6165fade1c9f40dae2543cbc89b