No need to recalc style before StyleResolver::computeFont.
StyleResolver::computeFont does not rely on the computed style of any
elements in the document tree. The updateLayoutTree call was there to
avoid an assert due to an unnecessary composed tree traversal for a
parent node. That traversal was removed by [1], so the updateLayoutTree
call should no longer be needed.

This change gets rid of synchronous style recalcs when FontFaceSet.load()
is called.

[1] https://codereview.chromium.org/1367743002/

BUG=536625

Review URL: https://codereview.chromium.org/1372823002

Cr-Original-Commit-Position: refs/heads/master@{#351025}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dc369a994337bd2c92948bf4d27d1549c063d6c2