TextAutosizer should handle frames without a FrameView.
During cross-process navigation, it is possible to create a LocalFrame
that doesn't have a FrameView, since the latter gets created at the time
a document commits. Since the commit in this case happens later in time,
the current code crashes as it expects the view to be there.

BUG=357747

Review URL: https://codereview.chromium.org/1374783005

Cr-Original-Commit-Position: refs/heads/master@{#351204}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6bdf645df32dbfa0ff4ae4925468519800f74827