[CSS Grid Layout] Flex tracks sizing alg must handle 0fr values
We don't allow 0 as flexible size value, which is not following current
specs; it just states that it must be a non-negative value. This patch
adds such change in the parser but some additional logic must be added
as well to handle 0 values during the flex tracks sizing algorithm.

The old algorithm didn't take 0 values into account, so there is the risk
of division by zero. Additionally, it was not handling fraction values
in the best way. The last versions of the spec changed this algorithm in
order to handle fraction values so that they don't cause exponential
grow of tracks using values bigger than 1.

This patch implements also such new algorithm, so we can deal not only
with 0 values, but managing fraction values properly.

BUG=520477

Review URL: https://codereview.chromium.org/1323053004

Cr-Original-Commit-Position: refs/heads/master@{#351301}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 993d4f780cb06a25b21613fead8fceca12dcdf6a