permissions: remove duplicates from request multiple at the Blink level
Propogating this information all the way to PermissionManager so it can be handled there
seems to be a waste of time and memory. It also adds a bunch of complexity to
all PermissionManager implementations.

Instead of this, handle all duplication at the Blink level so that chrome
can focus on handling requests.

BUG=516626

Review URL: https://codereview.chromium.org/1360023005

Cr-Original-Commit-Position: refs/heads/master@{#352072}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b3d2e697dc79112379f96f0f34cf26b83227a737