Don't deduct min-width from the available width for empty cells
If a column has cells, even if they are empty cells, then their full width needs
to be included in the width available for distribution to columns in the table.

This was an oversight when I introduced zero-width empty cells.

BUG=534830

Review URL: https://codereview.chromium.org/1368243002

Cr-Original-Commit-Position: refs/heads/master@{#352253}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5693a8522953e3511d31fd915716628b202d56f7