input[type=text]: Add a workaround to fix vertical position of Noto Sans CJK 1.004.
If an input element has the following style,
 - line-height: Npx
 - FontMetrics::height(): M (where M > N)

|line-height| for innerEditor was removed, and the innerEditor
and the input content box were aligned at the top edge.  So, the
bottom part of the text was clipped if |height| is specified and
it's smaller than FontMetrics::height().

This behavior is required because all browsers ignore
|line-height| if |height| is not specified.

Typically an error of author style causes this
situation. However, with Noto Sans 1.004, FontMetrics::height()
is larger than font-size, and the bottom part of text was clipped
even if <input style="font-size:16px; line-height:16px;
height:16px;">.  Authors don't expect this behavior.

This CL updates the code so that we don't remove |line-height|
from innerEditor if |height| is specified.  This centers
innerEditor vertically.

BUG=519331

Review URL: https://codereview.chromium.org/1387783003

Cr-Original-Commit-Position: refs/heads/master@{#352313}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bfa2f33f86fd947ce27dafbde21b3557b6d990fc