LayoutBlockFlow: rename paginationStrut to paginationStrutPropagatedFromChild.
While lines (RootInlineBox) store and keep the pagination struts that get
applied in front of them, blocks don't. So using the same term (simply
paginationStrut) in blocks and lines wasn't really appropriate [*]. This
paginationStrutPropagatedFromChild thing is really something that's only needed
during layout (see if children want to move a block to the next page column
and re-lay out if necessary). Ideally it shouldn't be stored in the objects at
all, but finding a way to avoid it proved hard. Stowing it in LayoutState was
one idea, but it turned out to be too complicated.

[*] The plan is to also store proper pagination struts on blocks (because
column balancing will benefit from it). This is a preparatory patch for that.

Did some light clean-up on my way, most notably: don't add in the old and
possibly stale pagination strut when estimating a logical top position in a
subsequent layout pass. The strut had typically been set to 0 at this point
anyway, but if it hasn't this code could possibly cause some mischief. In other
words, the code was either useless or harmful.

R=leviw@chromium.org, jchaffraix@chromium.org

Review URL: https://codereview.chromium.org/1381993002 .

Cr-Original-Commit-Position: refs/heads/master@{#352582}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 49591127bd411ffd9fc3c53009ac2c25da2e064f