Multicol: Add test that re-lays out a float and pushes it downwards at the same time.
The test passes, but the code seems somewhat brittle, so better make sure that
it keeps working.

Inline-level layout of floats is weird. LayoutBlockFlow::layoutInlineChildren()
lays out the float before its position is updated, and then marks the lines that
it thinks are affected as dirty (it may find the wrong lines here, but I haven't
been able to make anything fail because of it). For pagination this
additionally means that the pagination strut may be wrong (since the float's
position hasn't been updated). Luckily, we get an opportunity to lay out again
in insertFloatingObject() (still at the old position, though), and yet another
opportunity in positionNewFloats() (this time at the updated position).

R=jchaffraix@chromium.org,leviw@chromium.org,robhogan@gmail.com

Review URL: https://codereview.chromium.org/1375913003

Cr-Original-Commit-Position: refs/heads/master@{#352587}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 011ad4bb05558061cb4c8caf0204b552c3f9d77d