setLoadFinishTime on ImageResource for ImageDocument.
The loadFinishTime was never set on ImageResource used for the generated
image element inside an ImageDocument. We discovered this checking for
that time stamp in an internal Opera project.

Set the finish time of the ImageResource to that of the document (which
is the actual image).

Review URL: https://codereview.chromium.org/1380163006

Cr-Original-Commit-Position: refs/heads/master@{#352730}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bf9fbf106a8360f176ed1f7de9f3b87146e11f35