Protect access to m_value.
AudioParam::m_value is written by the main thread and also read (and
written) by the audio thread.  Atomically access this.  We don't
want a trylock around the audio thread because it's not clear what the
timeline should do if the the lock cannot be acquired.

Change m_value from a double to a float so we can do atomic
operations.  (All layout tests pass with this change.)  Then rename
m_value to m_intrinsicValue because that's what it is and add a setter
and getter for it so we can control access to the variable.

BUG=508447
TEST=tsan test passes

Review URL: https://codereview.chromium.org/1341793003

Cr-Original-Commit-Position: refs/heads/master@{#352865}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cb8d787f477bc5064b44d495773ef7e7909669df