[css-grid] Include freeSpace in GridSizingData struct
During the layout process we keep the free space for rows and columns in two
variables that are passed to a few methods along with the GridSizingData
struct.  Clearly those two variables should be part of GridSizingData as
they're temporary values used just for the sake of the layout.

Review URL: https://codereview.chromium.org/1387063003

Cr-Original-Commit-Position: refs/heads/master@{#353035}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bb66edce67d251e8db3327f03d9b48acf3cf1a6b