Fix a UAF in the BackgroundHtmlParser's constructor
Turns out the HTMLDocumentParser can go away before the task to start
the BackgroundHTMLParser runs on the background thread.  This can lead
to a UAF when we try to clone the WebTaskRunner.  The fix is to clone
it when posting instead.

BUG=541056, 510398

Review URL: https://codereview.chromium.org/1391013004

Cr-Original-Commit-Position: refs/heads/master@{#353083}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 76d7303845116b1c08bd327c46c99e264b626ada