Forward assigning to style on HTMLElement, SVGElement and CSSStyleRule to style.cssText
Currently setting element.style doesn't update the inline style of that
element. This is inconsistent with CSSOM spec ->
https://drafts.csswg.org/cssom/#cssstylerule. This patch forwards style
assignment on HTMLElement, SVGElement and CSSStyleRule to style.cssText
by changing the relevant idl files. The patch ensures that if the style
attribute is set, changed or removed, the declarations must be
updated as appropriate. FF supports forwarding assigning to style on
HTMLElement and SVGElement but not on CSSStyleRule. IE doesn't support
forwarding assignment to style on either of the three.

This patch also changes the layout tests in order to reflect the patch.
The expectation for checking in an input element's width and height
attributes without a loaded image has also been changed to a failing
expectation. This is as setting the display to none causes the
element and its descendants to not generate boxes hence returning a
width of 0.

BUG=533691

Review URL: https://codereview.chromium.org/1384973004

Cr-Original-Commit-Position: refs/heads/master@{#353235}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d1a7b0cbd0a511459f0f6202dfe61e27aa47df46