[css-grid] Remove unneeded calls to compute(Content)LogicalWidth(Height)
In order to resolve a Length to a LayoutUnit we need to provide a maximum
value so that i.e. percentages are correctly computed. That maximum value
was computeLogicalWidth() for columns and computeContentLogicalHeight() for
rows. We were calling it for every single track with definite size instead
of computing it once and reusing it multiple times.

This provides some nice performance improvements:
  - 7.75% in /Layout/fixed-grid-lots-of-data
  - 9.15% in /Layout/fixed-grid-lots-of-stretched-data

Review URL: https://codereview.chromium.org/1390913004

Cr-Original-Commit-Position: refs/heads/master@{#353304}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d20aa9cbc1ce40e084987084d0c8c6532379ec89