Oilpan: always shrink tail-allocated backing storage.
When shrinking backing storage allocations, some care is taken not to
introduce excessive fragmentation and/or limit later expansion of the
same object. That logic does not need to apply if the backing store is
at the tail of the current heap chunk: unconditionally reset the
allocation point, if it is.

R=haraken
BUG=420515

Review URL: https://codereview.chromium.org/1393863004

Cr-Original-Commit-Position: refs/heads/master@{#353321}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: afa3e61596ebb5d648a6e9923de85534926d27fa