Set a URL for cacheinvalidation.
The current URL is 'n/a' which, on the chrome://credits page will
just take the user to the top of the page, which is somewhat
confusing. Since this is now in-source, set the URL to the
README.chromium file itself, since that now contains the most useful
information.

BUG=chromium:476190

Review URL: https://codereview.chromium.org/1390403004

Cr-Original-Commit-Position: refs/heads/master@{#353357}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cf7419e1d8e3a77a10e27ed1f318f20669c6e62e